# E-commerce_project/E-commerce-Auth/src/main/java/com/example/authenticate/auth/AuthenticationResponse.java
package com.example.authenticate.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;

}